requirements.txt
asgiref==3.2.7
Django==3.0.5
psycopg2==2.8.5
pytz==2019.3
sqlparse==0.3.1
djangorestframework==3.11.0
